Reworked UI for setting up clip editing.
