+++ src/HearThis/app.config
@@ -94,13 +94,13 @@
             <setting name="AllowDisplayOfShiftClipsMenu" serializeAs="String">
                 <value>False</value>
             </setting>
-            <setting name="RestartingToChangeColorScheme" serializeAs="String">
-                <value>False</value>
-            </setting>
             <setting name="CurrentMode" serializeAs="String">
                 <value>ReadAndRecord</value>
             </setting>
             <setting name="EnableCheckForProblemsViewInProtectedMode" serializeAs="String">
+                <value>False</value>
+            </setting>
+            <setting name="RestartingToChangeColorScheme" serializeAs="String">
                 <value>False</value>
             </setting>
             <setting name="ExternalClipEditorPath" serializeAs="String">
